Refactor: DRYer dashboard item selection
