+++ lotti/pubspec.yaml
@@ -1,7 +1,7 @@
 name: lotti
 description: A Smart Journal.
 publish_to: 'none'
-version: 0.5.27+528
+version: 0.5.28+529
 
 environment:
   sdk: ">=2.14.0 <3.0.0"
